+++ src/Obsidian-All.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.779
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29306.81
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "global", "global", "{5B331FD3-D493-46A9-928B-B786FC6F3103}"
 	ProjectSection(SolutionItems) = preProject
@@ -182,8 +182,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Obsidian.Networks.ObsidianX
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Obsidian.Networks.Obsidian", "Obsidian.Networks.Obsidian\Obsidian.Networks.Obsidian.csproj", "{FA7BF227-B6A4-44E6-A7C7-C58317A0E79B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Obsidian.OxD", "Obsidian.OxD\Obsidian.OxD.csproj", "{0A167E2A-2930-4AE0-99B2-770BF6937806}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Obsidian.OxDroid", "Obsidian.OxDroid\Obsidian.OxDroid.csproj", "{0DEDC29E-F933-43FF-976F-F222B095F53B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Obsidian.ObsidianD", "Obsidian.ObsidianD\Obsidian.ObsidianD.csproj", "{E3B96125-6F07-477F-A3F1-5C76CF960485}"
@@ -198,7 +196,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Stratis.Features.Diagnostic
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Stratis.Features.Collateral", "Stratis.Features.Collateral\Stratis.Features.Collateral.csproj", "{8DB71680-D10C-4DAD-B922-FFCAFF7BE997}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Obsidian.ObsidianxD", "Obsidian.ObsidianxD\Obsidian.ObsidianxD.csproj", "{6E933548-5DC1-45F5-AB83-807F8AEDECAD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Obsidian.Ox1D", "Obsidian.Ox1D\Obsidian.Ox1D.csproj", "{CE4301DD-6C04-46B2-84A7-EDFE3B0E5960}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Obsidian.OxD", "Obsidian.OxD\Obsidian.OxD.csproj", "{B9AFF646-7A6B-49BD-B535-F8A37F2495CE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -486,10 +486,6 @@ Global
 		{FA7BF227-B6A4-44E6-A7C7-C58317A0E79B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FA7BF227-B6A4-44E6-A7C7-C58317A0E79B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FA7BF227-B6A4-44E6-A7C7-C58317A0E79B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0A167E2A-2930-4AE0-99B2-770BF6937806}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0A167E2A-2930-4AE0-99B2-770BF6937806}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0A167E2A-2930-4AE0-99B2-770BF6937806}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0A167E2A-2930-4AE0-99B2-770BF6937806}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0DEDC29E-F933-43FF-976F-F222B095F53B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0DEDC29E-F933-43FF-976F-F222B095F53B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0DEDC29E-F933-43FF-976F-F222B095F53B}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
@@ -522,10 +518,14 @@ Global
 		{8DB71680-D10C-4DAD-B922-FFCAFF7BE997}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8DB71680-D10C-4DAD-B922-FFCAFF7BE997}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8DB71680-D10C-4DAD-B922-FFCAFF7BE997}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6E933548-5DC1-45F5-AB83-807F8AEDECAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6E933548-5DC1-45F5-AB83-807F8AEDECAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6E933548-5DC1-45F5-AB83-807F8AEDECAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6E933548-5DC1-45F5-AB83-807F8AEDECAD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CE4301DD-6C04-46B2-84A7-EDFE3B0E5960}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE4301DD-6C04-46B2-84A7-EDFE3B0E5960}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE4301DD-6C04-46B2-84A7-EDFE3B0E5960}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE4301DD-6C04-46B2-84A7-EDFE3B0E5960}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B9AFF646-7A6B-49BD-B535-F8A37F2495CE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B9AFF646-7A6B-49BD-B535-F8A37F2495CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B9AFF646-7A6B-49BD-B535-F8A37F2495CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B9AFF646-7A6B-49BD-B535-F8A37F2495CE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -591,14 +591,14 @@ Global
 		{46E66D04-8869-4861-846A-18600604B974} = {4208F272-5766-46B4-8648-B5B5666166C5}
 		{921D2E58-AE2B-418C-B09D-D8D0EE98C0EA} = {4208F272-5766-46B4-8648-B5B5666166C5}
 		{FA7BF227-B6A4-44E6-A7C7-C58317A0E79B} = {E4693985-DDF7-4CF3-887B-0291135F9499}
-		{0A167E2A-2930-4AE0-99B2-770BF6937806} = {4208F272-5766-46B4-8648-B5B5666166C5}
 		{0DEDC29E-F933-43FF-976F-F222B095F53B} = {4208F272-5766-46B4-8648-B5B5666166C5}
 		{E3B96125-6F07-477F-A3F1-5C76CF960485} = {E4693985-DDF7-4CF3-887B-0291135F9499}
 		{1685ED2A-97A5-446D-9A79-3DD2BC4EC579} = {E4693985-DDF7-4CF3-887B-0291135F9499}
 		{C8D81134-8D49-4E68-8096-624312D8CA52} = {1B9A916F-DDAC-4675-B424-EDEDC1A58231}
 		{508D9F83-F30E-423C-A97A-EDE088CDB795} = {15D29FFD-6142-4DC5-AFFD-10BA0CA55C45}
 		{8DB71680-D10C-4DAD-B922-FFCAFF7BE997} = {15D29FFD-6142-4DC5-AFFD-10BA0CA55C45}
-		{6E933548-5DC1-45F5-AB83-807F8AEDECAD} = {4208F272-5766-46B4-8648-B5B5666166C5}
+		{CE4301DD-6C04-46B2-84A7-EDFE3B0E5960} = {E4693985-DDF7-4CF3-887B-0291135F9499}
+		{B9AFF646-7A6B-49BD-B535-F8A37F2495CE} = {E4693985-DDF7-4CF3-887B-0291135F9499}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6C780ABA-5872-4B83-AD3F-A5BD423AD907}
