--- src/Obsidian.ObsidianxD/Program.cs
@@ -1,51 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using NBitcoin;
-using NBitcoin.Protocol;
-using Obsidian.Networks.ObsidianX;
-using Stratis.Bitcoin.Builder;
-using Stratis.Bitcoin.Configuration;
-using Stratis.Bitcoin.Features.Api;
-using Stratis.Bitcoin.Features.BlockStore;
-using Stratis.Bitcoin.Features.ColdStaking;
-using Stratis.Bitcoin.Features.Consensus;
-using Stratis.Bitcoin.Features.MemoryPool;
-using Stratis.Bitcoin.Features.Miner;
-using Stratis.Bitcoin.Features.RPC;
-using Stratis.Bitcoin.Utilities;
-
-namespace Obsidian.ObsidianxD
-{
-    public class Program
-    {
-        public static async Task Main(string[] args)
-        {
-            PosBlockHeader.CustomPoWHash = ObsidianXHash.GetObsidianXPoWHash;
-
-            try
-            {
-                var nodeSettings = new NodeSettings(
-                    networksSelector: ObsidianXNetworksSelector.Obsidian, 
-                    protocolVersion: ProtocolVersion.PROVEN_HEADER_VERSION, 
-                    agent: $"obsidianx", 
-                    args: args);
-
-                var builder = new FullNodeBuilder()
-                    .UseNodeSettings(nodeSettings)
-                    .UseBlockStore()
-                    .UsePosConsensus()
-                    .UseMempool()
-                    .UseColdStakingWallet()
-                    .AddPowPosMining()
-                    .AddRPC()
-                    .UseApi();
-
-                await builder.Build().RunAsync();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(@"There was a problem initializing the node. Details: '{0}'", ex.Message);
-            }
-        }
-    }
-}
