+++ src/Obsidian.Ox1D/NLog.config
