+++ src/Obsidian.Ox1D/Properties/launchSettings.json
@@ -1,0 +1,12 @@
+{
+  "profiles": {
+    "Obsidian.OxD": {
+      "commandName": "Project",
+      "commandLineArgs": "-addnode=165.22.90.248",
+      "environmentVariables": {
+        "deleteme2": "cli getinfo",
+        "deleteme": "-port=56666 -addnode=165.22.90.248"
+      }
+    }
+  }
+}
