+++ src/Obsidian.Ox1D/Temp/Miner.cs
