+++ src/Obsidian.Ox1D/Temp/ODXMiner.cs
