+++ src/Obsidian.OxD/Program.cs
@@ -1,14 +1,8 @@
 ﻿using System;
-using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using NBitcoin;
 using NBitcoin.Protocol;
-using Obsidian.Features.X1Wallet;
-using Obsidian.Features.X1Wallet.SecureApi;
 using Obsidian.Networks.ObsidianX;
-using Obsidian.OxD.Api;
-using Obsidian.OxD.Cli;
 using Stratis.Bitcoin.Builder;
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Features.Api;
@@ -20,92 +14,38 @@ using Stratis.Bitcoin.Features.Miner;
 using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Utilities;
 
-namespace Obsidian.OxD
+namespace Obsidian.ObsidianxD
 {
-    public static class Program
+    public class Program
     {
-        public static void Main(string[] args)
-        {
-            if (args != null && args.Length > 0 && args[0] == "cli")
-            {
-                var argList = args.ToList();
-                argList.RemoveAt(0);
-                CliTool.CliMain(argList.ToArray());
-            }
-            else
-            {
-                MainAsync(args).Wait();
-            }
-        }
-
-        /// <summary> Starts the fullnode asyncronously.</summary>
-        /// <remarks>To run as gateway node use the args -gateway=1 -whitelist=[trusted-QT-ip] addnode=[trusted-QT-ip],
-        /// e.g. -gateway=1 -whitelist=104.45.21.229 addnode=104.45.21.229 -port=56666
-        /// Use arg -maxblkmem=2 on a low memory VPS.
-        /// </remarks>
-        /// <param Command="args">args</param>
-        /// <returns><placeholder>A <see cref="Task"/> representing the asynchronous operation.</placeholder></returns>
-        static async Task MainAsync(string[] args)
+        public static async Task Main(string[] args)
         {
             PosBlockHeader.CustomPoWHash = ObsidianXHash.GetObsidianXPoWHash;
 
             try
             {
-                var nodeSettings = new NodeSettings(networksSelector: ObsidianXNetworksSelector.Obsidian,
-                    protocolVersion: ProtocolVersion.PROVEN_HEADER_VERSION, agent: $"{GetName()}, Stratis ", args: args)
-                {
-                    MinProtocolVersion = ProtocolVersion.PROVEN_HEADER_VERSION
-                };
-
-                var useHDWallet = args.Contains("-hdwallet");
+                var nodeSettings = new NodeSettings(
+                    networksSelector: ObsidianXNetworksSelector.Obsidian, 
+                    protocolVersion: ProtocolVersion.PROVEN_HEADER_VERSION, 
+                    agent: $"obsidianx", 
+                    args: args);
 
                 var builder = new FullNodeBuilder()
-                            .UseNodeSettings(nodeSettings)
-                            .UseBlockStore()
-                            .UsePosConsensus()
-                            .UseMempool();
+                    .UseNodeSettings(nodeSettings)
+                    .UseBlockStore()
+                    .UsePosConsensus()
+                    .UseMempool()
+                    .UseColdStakingWallet()
+                    .AddPowPosMining()
+                    .AddRPC()
+                    .UseApi();
 
-
-                if (useHDWallet)
-                {
-                    builder
-                        .AddPowPosMining()
-                        .AddRPC()
-                        .UseColdStakingWallet()
-                        .UseApi();
-                }
-                else
-                {
-                    builder.UseX1Wallet()
-                        .UseX1WalletApi()
-                        .UseSecureApiHost();
-                }
-
-                var node = builder.Build();
-
-#if DEBUG
-                //_ = Task.Run(async () =>
-                //  {
-                //      await Task.Delay(15000);
-                //      TestBench.Run((FullNode)node);  // start mining to the wallet
-                //  });
-#endif
-
-                await node.RunAsync();
+                await builder.Build().RunAsync();
             }
             catch (Exception ex)
             {
                 Console.WriteLine(@"There was a problem initializing the node. Details: '{0}'", ex.Message);
             }
-        }
-
-        static string GetName()
-        {
-#if DEBUG
-            return $"oxd {Assembly.GetEntryAssembly()?.GetName().Version} (d)";
-#else
-			return $"oxd {Assembly.GetEntryAssembly()?.GetName().Version} (r)";
-#endif
         }
     }
 }
